Rename 'exceptions' files to 'errors'
