+++ couchbase_columnar/common/errors.py
